requirements.txt
# Version 8

jinja2
# https://stackoverflow.com/questions/1960516/python-json-serialize-a-decimal-object
simplejson>=3.18.4
pik_django_utils @ git+https://github.com/pik-software/pik-django-utils.git@add-drf

# tests
pytest
pytest-cov
prospector
mypy
